Add a global 'Cache' collection for generic caches
Added a global cache collection that currently stores a single global cache for persistent storage of model data. It also adds a `persist` method to the Movies model to be used to store persistent metadata related to that movie in a database. It will be loaded into the models attributes when the model is instantiated.